[C#] Checkpoint, Recovery, RecordInfo updates (#588)
* Removed WAIT_PENDING

* * Use only 1 bit for newVersion in RecordInfo, for CPR. Semantics are that sessions in v+1 will create records in fuzzy region with this bit set. Recovery only scans fuzzy region to elide such records.
* Add 1 Filler bit for future use with fillers in records
* Increase read lock bits to 6 (64 parallel readers)
* Use dirty bit for incremental checkpoints

* fix ycsb to not delete data file if we are using --recover

* nits

* Fix epoch re-entry violation

* Move to long versions, for KV checkpoints.

* Shuffle around stuff for better cache alignment in benchmark.

* Updates based on review

+++ cs/src/core/Epochs/LightEpoch.cs
@@ -474,9 +474,9 @@ namespace FASTER.core
         /// <param name="version">Version</param>
         /// <returns></returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void Mark(int markerIdx, int version)
+        public void Mark(int markerIdx, long version)
         {
-            (*(tableAligned + threadEntryIndex)).markers[markerIdx] = version;
+            (*(tableAligned + threadEntryIndex)).markers[markerIdx] = (int)version;
         }
 
         /// <summary>
@@ -487,7 +487,7 @@ namespace FASTER.core
         /// <param name="version">Version</param>
         /// <returns>Whether complete</returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public bool CheckIsComplete(int markerIdx, int version)
+        public bool CheckIsComplete(int markerIdx, long version)
         {
             // check if all threads have reported complete
             for (int index = 1; index <= kTableSize; ++index)
@@ -496,7 +496,7 @@ namespace FASTER.core
                 int fc_version = (*(tableAligned + index)).markers[markerIdx];
                 if (0 != entry_epoch)
                 {
-                    if (fc_version != version && entry_epoch < int.MaxValue)
+                    if ((fc_version != (int)version) && (entry_epoch < int.MaxValue))
                     {
                         return false;
                     }
